Pass along project changes through shared workspace
Circle does not currently have the concept of an optional job, so my
idea to have the "Project A" job depend upon a "Project A Changed" job
won't work, because if "Project A Changed" fails, the whole build fails.

So for now, we're going to pass along some sentinel files that represent
whether our projects changed, and reference those in our dependent jobs.